+++ backend/src/main/java/com/woowacourse/levellog/interviewquestion/dto/InterviewQuestionDetailDto.java
@@ -1,6 +1,6 @@
-package com.woowacourse.levellog.interview_question.dto;
+package com.woowacourse.levellog.interviewquestion.dto;
 
-import com.woowacourse.levellog.interview_question.domain.InterviewQuestion;
+import com.woowacourse.levellog.interviewquestion.domain.InterviewQuestion;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
