+++ backend/src/test/java/com/woowacourse/levellog/acceptance/InterviewQuestionAcceptanceTest.java
@@ -8,7 +8,7 @@ import static org.hamcrest.Matchers.notNullValue;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 import com.woowacourse.levellog.fixture.RestAssuredResponse;
-import com.woowacourse.levellog.interview_question.dto.InterviewQuestionDto;
+import com.woowacourse.levellog.interviewquestion.dto.InterviewQuestionDto;
 import com.woowacourse.levellog.levellog.dto.LevellogDto;
 import com.woowacourse.levellog.team.dto.ParticipantIdsDto;
 import com.woowacourse.levellog.team.dto.TeamCreateDto;
@@ -35,7 +35,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
     void save() {
         // given
         final RestAssuredResponse hostLoginResponse = login("페퍼");
-        final Long pepperId = hostLoginResponse.getMemberId();
+        hostLoginResponse.getMemberId();
         final String pepperToken = hostLoginResponse.getToken();
 
         final RestAssuredResponse romaLoginResponse = login("로마");
@@ -43,7 +43,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
         final String romaToken = romaLoginResponse.getToken();
 
         final TeamCreateDto teamCreateDto = new TeamCreateDto("롬펲 인터뷰", "트랙룸", LocalDateTime.now().plusDays(3),
-                new ParticipantIdsDto(List.of(pepperId, romaId)));
+                new ParticipantIdsDto(List.of(romaId)));
         final String teamId = post("/api/teams", pepperToken, teamCreateDto).getTeamId();
 
         final LevellogDto levellogRequest = LevellogDto.from("페퍼의 레벨로그");
@@ -78,7 +78,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
     void findAll() {
         // given
         final RestAssuredResponse hostLoginResponse = login("페퍼");
-        final Long pepperId = hostLoginResponse.getMemberId();
+        hostLoginResponse.getMemberId();
         final String pepperToken = hostLoginResponse.getToken();
 
         final RestAssuredResponse romaLoginResponse = login("로마");
@@ -86,7 +86,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
         final String romaToken = romaLoginResponse.getToken();
 
         final TeamCreateDto teamCreateDto = new TeamCreateDto("롬펲 인터뷰", "트랙룸", LocalDateTime.now().plusDays(3),
-                new ParticipantIdsDto(List.of(pepperId, romaId)));
+                new ParticipantIdsDto(List.of(romaId)));
         final String teamId = post("/api/teams", pepperToken, teamCreateDto).getTeamId();
 
         final LevellogDto levellogRequest = LevellogDto.from("페퍼의 레벨로그");
@@ -122,7 +122,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
     void update() {
         // given
         final RestAssuredResponse hostLoginResponse = login("페퍼");
-        final Long pepperId = hostLoginResponse.getMemberId();
+        hostLoginResponse.getMemberId();
         final String pepperToken = hostLoginResponse.getToken();
 
         final RestAssuredResponse romaLoginResponse = login("로마");
@@ -130,7 +130,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
         final String romaToken = romaLoginResponse.getToken();
 
         final TeamCreateDto teamCreateDto = new TeamCreateDto("롬펲 인터뷰", "트랙룸", LocalDateTime.now().plusDays(3),
-                new ParticipantIdsDto(List.of(pepperId, romaId)));
+                new ParticipantIdsDto(List.of(romaId)));
         final String teamId = post("/api/teams", pepperToken, teamCreateDto).getTeamId();
 
         final LevellogDto levellogRequest = LevellogDto.from("페퍼의 레벨로그");
@@ -170,7 +170,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
     void deleteById() {
         // given
         final RestAssuredResponse hostLoginResponse = login("페퍼");
-        final Long pepperId = hostLoginResponse.getMemberId();
+        hostLoginResponse.getMemberId();
         final String pepperToken = hostLoginResponse.getToken();
 
         final RestAssuredResponse romaLoginResponse = login("로마");
@@ -178,7 +178,7 @@ class InterviewQuestionAcceptanceTest extends AcceptanceTest {
         final String romaToken = romaLoginResponse.getToken();
 
         final TeamCreateDto teamCreateDto = new TeamCreateDto("롬펲 인터뷰", "트랙룸", LocalDateTime.now().plusDays(3),
-                new ParticipantIdsDto(List.of(pepperId, romaId)));
+                new ParticipantIdsDto(List.of(romaId)));
         final String teamId = post("/api/teams", pepperToken, teamCreateDto).getTeamId();
 
         final LevellogDto levellogRequest = LevellogDto.from("페퍼의 레벨로그");
