+++ backend/src/test/java/com/woowacourse/levellog/application/ServiceTest.java
@@ -6,8 +6,8 @@ import com.woowacourse.levellog.authentication.support.JwtTokenProvider;
 import com.woowacourse.levellog.config.TestAuthenticationConfig;
 import com.woowacourse.levellog.feedback.application.FeedbackService;
 import com.woowacourse.levellog.feedback.domain.FeedbackRepository;
-import com.woowacourse.levellog.interview_question.application.InterviewQuestionService;
-import com.woowacourse.levellog.interview_question.domain.InterviewQuestionRepository;
+import com.woowacourse.levellog.interviewquestion.application.InterviewQuestionService;
+import com.woowacourse.levellog.interviewquestion.domain.InterviewQuestionRepository;
 import com.woowacourse.levellog.levellog.application.LevellogService;
 import com.woowacourse.levellog.levellog.domain.LevellogRepository;
 import com.woowacourse.levellog.member.application.MemberService;
