+++ backend/src/test/java/com/woowacourse/levellog/domain/InterviewQuestionRepositoryTest.java
@@ -3,8 +3,8 @@ package com.woowacourse.levellog.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.woowacourse.levellog.common.config.JpaConfig;
-import com.woowacourse.levellog.interview_question.domain.InterviewQuestion;
-import com.woowacourse.levellog.interview_question.domain.InterviewQuestionRepository;
+import com.woowacourse.levellog.interviewquestion.domain.InterviewQuestion;
+import com.woowacourse.levellog.interviewquestion.domain.InterviewQuestionRepository;
 import com.woowacourse.levellog.levellog.domain.Levellog;
 import com.woowacourse.levellog.levellog.domain.LevellogRepository;
 import com.woowacourse.levellog.member.domain.Member;
