+++ backend/src/test/java/com/woowacourse/levellog/presentation/ControllerTest.java
@@ -13,8 +13,8 @@ import com.woowacourse.levellog.authentication.presentation.OAuthController;
 import com.woowacourse.levellog.authentication.support.JwtTokenProvider;
 import com.woowacourse.levellog.feedback.application.FeedbackService;
 import com.woowacourse.levellog.feedback.presentation.FeedbackController;
-import com.woowacourse.levellog.interview_question.application.InterviewQuestionService;
-import com.woowacourse.levellog.interview_question.presentation.InterviewQuestionController;
+import com.woowacourse.levellog.interviewquestion.application.InterviewQuestionService;
+import com.woowacourse.levellog.interviewquestion.presentation.InterviewQuestionController;
 import com.woowacourse.levellog.levellog.application.LevellogService;
 import com.woowacourse.levellog.levellog.presentation.LevellogController;
 import com.woowacourse.levellog.member.application.MemberService;
